// src/api-serverless/src/generated/models/ApiCreateConnectionTransferResponse.ts
/**
 * 6529.io API
 * This is the API interface description. Brief terminology overview and an authentication example can be found at <a href=\"https://6529.io/about/api\">https://6529.io/about/api</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiCreateConnectionTransferResponse {
    'transfer_code': string;
    'expires_at': Date;
    'address': string;
    'role'?: string | null;
    'target_client_type': ApiCreateConnectionTransferResponseTargetClientTypeEnum;
    'deep_link_path': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "transfer_code",
            "baseName": "transfer_code",
            "type": "string",
            "format": ""
        },
        {
            "name": "expires_at",
            "baseName": "expires_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "address",
            "baseName": "address",
            "type": "string",
            "format": ""
        },
        {
            "name": "role",
            "baseName": "role",
            "type": "string",
            "format": ""
        },
        {
            "name": "target_client_type",
            "baseName": "target_client_type",
            "type": "ApiCreateConnectionTransferResponseTargetClientTypeEnum",
            "format": ""
        },
        {
            "name": "deep_link_path",
            "baseName": "deep_link_path",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCreateConnectionTransferResponse.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiCreateConnectionTransferResponseTargetClientTypeEnum {
    Native = 'native'
}