// src/api-serverless/src/generated/models/ApiRedeemConnectionTransferResponse.ts
/**
 * 6529.io API
 * This is the API interface description. Brief terminology overview and an authentication example can be found at <a href=\"https://6529.io/about/api\">https://6529.io/about/api</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiRedeemConnectionTransferResponse {
    'address': string;
    'role'?: string | null;
    'access_token': string;
    'access_token_expires_at': Date;
    'native_refresh_token': string;
    'refresh_token_expires_at': Date;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "address",
            "baseName": "address",
            "type": "string",
            "format": ""
        },
        {
            "name": "role",
            "baseName": "role",
            "type": "string",
            "format": ""
        },
        {
            "name": "access_token",
            "baseName": "access_token",
            "type": "string",
            "format": ""
        },
        {
            "name": "access_token_expires_at",
            "baseName": "access_token_expires_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "native_refresh_token",
            "baseName": "native_refresh_token",
            "type": "string",
            "format": ""
        },
        {
            "name": "refresh_token_expires_at",
            "baseName": "refresh_token_expires_at",
            "type": "Date",
            "format": "date-time"
        }    ];

    static getAttributeTypeMap() {
        return ApiRedeemConnectionTransferResponse.attributeTypeMap;
    }

    public constructor() {
    }
}