// src/api-serverless/src/generated/models/ApiSessionWebResponse.ts
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ApiSessionWebResponse {
    'address': string;
    'role'?: string | null;
    'access_token': string;
    'access_token_expires_at': Date;
    'client_type': ApiSessionWebResponseClientTypeEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "address",
            "baseName": "address",
            "type": "string",
            "format": ""
        },
        {
            "name": "role",
            "baseName": "role",
            "type": "string",
            "format": ""
        },
        {
            "name": "access_token",
            "baseName": "access_token",
            "type": "string",
            "format": ""
        },
        {
            "name": "access_token_expires_at",
            "baseName": "access_token_expires_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "client_type",
            "baseName": "client_type",
            "type": "ApiSessionWebResponseClientTypeEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiSessionWebResponse.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiSessionWebResponseClientTypeEnum {
    Web = 'web'
}